chore: Fix updating the Gemfile.lock.
`bundle install` failed because we are in deployment mode and the
Gemfile.lock cannot be updated.

```
You are trying to install in deployment mode after changing
your Gemfile. Run `bundle install` elsewhere and add the
updated Gemfile.lock to version control.

If this is a development machine, remove the
/home/runner/work/client-sdk-ruby/client-sdk-ruby/Gemfile freeze
by running `bundle config unset deployment`.
```