chore: Fix release-please PR tests
The PR tests will fail on a Release Please PR because they run in
bundler development mode. The Release Please PR bumps the version
number which requires bundler to update the Gemfile.lock; it will
not do this in deployment mode.

Also, we want all release changes to be in the PR.