chore: switch cache scalar and control methods to a new pattern (#145)
Create a new integration directory in spec containing new tests for the
cache client scalar and control functions. They use a common helper that
creates a cache before the suite and deletes it after. Future functions
can be tested here or we can make new files for the different
collections.

Remove next_token from list_caches, as it is unused. This simplifies the
'caches' method. Should we remove it?

Rename auth_token to api_key and TEST_AUTH_TOKEN to MOMENTO_API_KEY.

Default to 'ruby-test-cache' if TEST_CACHE_NAME is not specified.

Remove the flag that prevented network tests from running.

Remove the now redundant live_spec tests.