feat: support UTF-8 cache names instead of just ASCII
Update the validate_cache_name function to return a UTF-8 encoded
version of the cache name. If it already is UTF-8, this won't do
anything. If it contains bytes that can't be encoded in UTF-8, it raises
an error.

Use the result of validate_cache_name in the cache functions.